Convert TradeLogs to Strings on write
This is totally untested as I wasn't able to pull the PluginBase dependency easily, but should fix Trophonix/TradePlus#60 and fix Trophonix/TradePlus#62 if I didn't mess anything up.

I believe the root of the stackoverflow has something to do with ItemStack not being serializable (`org.bukkit.configuration.serialization.ConfigurationSerializable` != `java.io.Serializable`), so I've switched to using strings.

I'm making an assumption in this patch that there shouldn't be a reason we would need to restore a TradeLog from what is written to the file since that seems like overkill for basic logging. If that's a bad assumption, we should implement a better TypeAdapter that will handle the serialization of ItemStacks.